add: create dedicated Video Service for helper functions
